Update `el-get-sources' properly.
When `bundle-el-get' is called inside `bundle' macro (which means
the callsite appears in the :after script), `el-get' called by
`bundle-el-get' restores the value of `el-get-sources'. This is not
expected behavior since `bundle-el-get' tries to update `el-get-sources'.

This change makes sure that the value of `el-get-sources' is updated
after the nested call of `bundle-el-get'.